// src/app/page/master/map-bahasa-to-negara/map-bahasa-to-negara.component.interface.ts
import {EventEmitter} from '@angular/core';

export interface PegawaiSkTidakAbsen {
	kdProduk?: string;
	kdEvent?: number;
	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}